requirements.txt
openai
sarvamai
numpy
sounddevice
webrtcvad
soundfile
streamlit
python-dotenv
